pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deid"
version = "0.1.0"
description = "CLI tool for deidentifying financial CSV files"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pandas>=1.5",
    "gspread>=5.0",
    "google-auth-oauthlib>=1.0",
    "google-auth>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
deid = "deid.cli:cli"

[tool.setuptools.packages.find]
include = ["deid*"]
